Bumped spdlog

+++ src/Eis/Rendering/Objects/Texture.cpp
@@ -20,7 +20,7 @@ namespace Eis
 			return CreateRef<OpenGLES2Texture2D>(path);
 
 		default:
-			EIS_CORE_ASSERT(false, "Invalid graphics API! ({0})", Renderer2D::GetAPI());
+			EIS_CORE_ASSERT(false, "Invalid graphics API! ({0})", (uint8_t)Renderer2D::GetAPI());
 			return nullptr;
 		}
 	}
@@ -36,7 +36,7 @@ namespace Eis
 			return CreateRef<OpenGLES2Texture2D>(width, height);
 
 		default:
-			EIS_CORE_ASSERT(false, "Invalid graphics API! ({0})", Renderer2D::GetAPI());
+			EIS_CORE_ASSERT(false, "Invalid graphics API! ({0})", (uint8_t)Renderer2D::GetAPI());
 			return nullptr;
 		}
 	}
@@ -52,7 +52,7 @@ namespace Eis
 			return CreateRef<OpenGLES2Texture2D>(image);
 
 		default:
-			EIS_CORE_ASSERT(false, "Invalid graphics API! ({0})", Renderer2D::GetAPI());
+			EIS_CORE_ASSERT(false, "Invalid graphics API! ({0})", (uint8_t)Renderer2D::GetAPI());
 			return nullptr;
 		}
 	}
